requirements-jetson.txt
tflite_runtime-2.5.0-cp36-cp36m-linux_aarch64.whl
Pillow==9.0.0
numpy==1.17.4
